refactor(cmd): add cobra cli libary
