+++ AGENTS.md
@@ -45,12 +45,12 @@ The core functionality involves serving WebP images to devices, generated by ren
 1.  **Dependencies:** Ensure Go 1.26+ and C build tools (gcc) are installed.
 2.  **Run with `go run` (for development):**
     ```bash
-    go run ./cmd/server -db data/tronbyt.db -data .
+    go run ./cmd/server --db data/tronbyt.db --data .
     ```
 3.  **Build and Run (for production/deployment):**
     ```bash
     go build -o tronbyt-server ./cmd/server
-    ./tronbyt-server -db data/tronbyt.db -data .
+    ./tronbyt-server --db data/tronbyt.db --data .
     ```
 
 ### Testing
