+++ cmd/server/boot/privileges_unix.go
@@ -1,0 +1,21 @@
+//go:build !windows
+
+package boot
+
+import (
+	"fmt"
+	"syscall"
+)
+
+func dropPrivileges(uid, gid int) error {
+	if err := syscall.Setgroups([]int{gid}); err != nil {
+		return fmt.Errorf("failed to set groups: %w", err)
+	}
+	if err := syscall.Setgid(gid); err != nil {
+		return fmt.Errorf("failed to set gid: %w", err)
+	}
+	if err := syscall.Setuid(uid); err != nil {
+		return fmt.Errorf("failed to set uid: %w", err)
+	}
+	return nil
+}
