+++ internal/config/flags.go
@@ -1,0 +1,6 @@
+package config
+
+const (
+	FlagDB   = "db"
+	FlagData = "data"
+)
